Simplify: extract shared AR cursor module, minor cleanups
- Extract ActiveRecordCursor module shared by both AR enumerators
  (eliminates 4 duplicated private methods)
- Cache ordered scope in build_scope to avoid rebuilding Arel per batch
- @collection_enum ivar → local var (only used within step block)
- check_throttle: use find idiom instead of each + early return
