+++ app/models/maintenance_tasks/active_record_record_enumerator.rb
@@ -4,6 +4,7 @@ module MaintenanceTasks
   # @api private
   class ActiveRecordRecordEnumerator
     include Enumerable
+    include ActiveRecordCursor
 
     DEFAULT_BATCH_SIZE = 100
 
@@ -25,8 +26,7 @@ module MaintenanceTasks
 
       cursor = @cursor
       loop do
-        scope = build_scope(cursor)
-        batch = scope.limit(@batch_size).to_a
+        batch = build_scope(cursor).limit(@batch_size).to_a
         break if batch.empty?
 
         batch.each do |record|
@@ -34,42 +34,6 @@ module MaintenanceTasks
           yield record, cursor
         end
       end
-    end
-
-    private
-
-    def build_scope(cursor)
-      scope = @relation.reorder(@columns.map { |col| arel_table[col].asc })
-      return scope if cursor.nil?
-
-      cursor_values = Array(cursor)
-      scope.where(cursor_conditions(cursor_values))
-    end
-
-    def cursor_conditions(cursor_values)
-      conditions = @columns.each_index.map do |i|
-        eq_clauses = @columns[0...i].map.with_index do |col, j|
-          arel_table[col].eq(cursor_values[j])
-        end
-        gt_clause = arel_table[@columns[i]].gt(cursor_values[i])
-
-        if eq_clauses.empty?
-          gt_clause
-        else
-          eq_clauses.reduce(:and).and(gt_clause)
-        end
-      end
-
-      conditions.reduce(:or)
-    end
-
-    def extract_cursor(record)
-      values = @columns.map { |col| record.public_send(col) }
-      values.size == 1 ? values.first : values
-    end
-
-    def arel_table
-      @relation.klass.arel_table
     end
   end
 end
